解决长度丢一个字节； 红外电源打开；

+++ User/misc.h
@@ -4,7 +4,7 @@
 #include "includes.h"
 
 void delay_us(u32 us);
-void DecToBCD(u32 Dec, u8 *pBCD, u8 len);
+void DecToBCD(u16 Dec, u8 *pBCD, u8 len);
 float CalculateTemperature(float dat,const u16 Rp,const u16 Bx);
 
 #endif
